requirements.txt
torch==2.5.1+cu118
torchvision==0.20.1+cu118
torchaudio==2.5.1+cu118
numpy==1.26.4
pandas==2.3.1
scikit-learn==1.7.1
scikit-image==0.25.0
scipy==1.15.3
matplotlib==3.10.5
seaborn==0.13.2
opencv-python==4.1.2.30
opencv-contrib-python==4.12.0.88
Pillow==11.3.0
tqdm==4.60.0
PyYAML==6.0.2
h5py==2.10.0
openpyxl==3.1.2
scikit-optimize==0.10.2
tensorboard==2.13.0
torchsummary==1.5.1
jittor==1.3.10.0
transformers==4.46.3
timm==0.9.2